authorization

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ class PostController extends Controller
 {
     public function index()
     {
-        $posts = Post::with((['user','likes']))->paginate(20);
+        $posts = Post::latest()->with((['user','likes']))->paginate(20);
         return view('posts.index',
         [
             'posts' =>$posts
@@ -36,6 +36,12 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        // if (!$post->ownedBy(auth()->user()))
+        // {
+        //     dd('no');
+        // }
+
+        $this->authorize('delete', $post);
         $post->delete();
         return back();
     }
